+++ internal/service/notification_service_scheduler_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/temirov/pinguin/pkg/model"
+	"github.com/temirov/pinguin/internal/model"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 	"log/slog"